requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
asyncpg
python-dotenv 
aiohttp 
openai
psutil 